requirements.txt
pandas>=1.3.0
numpy>=1.21.0
requests>=2.26.0
streamlit>=1.0.0
openai>=0.27.0
python-dotenv>=0.19.0
beautifulsoup4>=4.10.0
lxml>=4.6.3
plotly>=5.3.1
matplotlib>=3.4.3
